Refactor serve
